pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.121.2",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "uvicorn[standard]>=0.38.0",
]
